# services/contact/internal/delivery/http/swagger/docs/swagger.yaml
basePath: /
definitions:
...
  group.GroupList:
    properties:
      limit:
        default: 10
        description: Количество записей
        example: 10
        minimum: 0
        type: integer
      list:
        description: Список групп
        items:
          $ref: '#/definitions/group.GroupResponse'
        minItems: 0
        type: array
      offset:
        default: 0
        description: Смещение при получении записей
        example: 20
        minimum: 0
        type: integer
      total:
        default: 0
        description: Всего
        example: 10
        minimum: 0
        type: integer
    type: object
  group.GroupResponse:
    properties:
      contactsAmount:
        default: 10
        description: Кол-во контактов в группе
        minimum: 0
        type: integer
      createdAt:
        description: Дата создания группы
        type: string
      description:
        description: Описание
        example: Описание группы
        maxLength: 1000
        type: string
      id:
        description: Идентификатор группы
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        type: string
      modifiedAt:
        description: Дата последнего изменения группы
        type: string
      name:
        description: Название группы
        example: Название группы
        maxLength: 100
        type: string
    required:
    - createdAt
    - id
    - modifiedAt
    - name
    type: object
  group.ShortGroup:
    properties:
      description:
        description: Описание
        example: Описание группы
        maxLength: 1000
        type: string
      name:
        description: Название группы
        example: Название группы
        maxLength: 100
        type: string
    required:
    - name
    type: object
  http.ErrorResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      id:
        type: string
      info: {}
      message:
        type: string
    type: object
info:
  contact:
    email: kolyadkons@gmail.com
    name: API Support
  description: contact service on clean architecture
  license:
    name: kolyadkons
  title: slurm contact service on clean architecture
  version: "1.0"
paths:
...
  /groups/:
    get:
      consumes:
      - application/json
      description: Метод позволяет получить список групп.
      parameters:
      - default: 10
        description: Количество записей
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение при получении записей
        in: query
        name: offset
        type: integer
      - default: name
        description: Сортировка по полю
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/group.GroupList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
        "404":
          description: 404 Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Метод позволяет получить список групп.
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Метод позволяет создать группу контактов.
      parameters:
      - description: Данные по группе
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/group.ShortGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/group.GroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
        "404":
          description: 404 Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Метод позволяет создать группу контактов.
      tags:
      - groups
  /groups/{id}:
    delete:
      consumes:
      - application/json
      description: Метод позволяет удалить группу.
      parameters:
      - description: Идентификатор группы
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
        "404":
          description: 404 Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Метод позволяет удалить группу.
      tags:
      - groups
    get:
      consumes:
      - application/json
      description: Метод позволяет получить данные по группе.
      parameters:
      - description: Идентификатор группы контактов
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/group.GroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
        "404":
          description: 404 Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Метод позволяет получить данные по группе.
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Метод позволяет обновить данные группы.
      parameters:
      - description: Идентификатор группы
        in: path
        name: id
        required: true
        type: string
      - description: Данные по группе
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/group.ShortGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/group.GroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
        "404":
          description: 404 Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Метод позволяет обновить данные группы.
      tags:
      - groups
...
swagger: "2.0"